Fixed "suicider" Killed event handler
This was buggy in the default implementation (sometimes variable would not be `false` on `_city` where the bomber was killed). Fixed by passing in `_city` reference directly and setting object variable on the client owning the object (server), rather than the killed unit owner (could be server or HC).

EH registered with CBA_fnc_addBISEventHandler for convenience.